requirements.txt
Flask==2.1.1
gunicorn==20.1.0
python-dotenv==0.20.0
Flask-Minify==0.39
Flask-SQLAlchemy==2.5.1 
SQLAlchemy-Utils==0.38.2
Flask-Migrate==3.1.0
psycopg2-binary==2.9.3